requirement.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
